refactor(app): move pages into features
