+++ web/app/src/features/auth/index.tsx
@@ -1,0 +1,2 @@
+export { default as LoginPage } from "./login-page";
+export { default as RegisterPage } from "./register-page";
