+++ web/app/src/features/graphiql/index.tsx
@@ -1,0 +1,1 @@
+export { default as Page } from "./page";
