+++ web/app/src/features/projects/index.tsx
@@ -1,0 +1,1 @@
+export { default as Page } from "./page";
