+++ web/app/src/features/user/index.tsx
@@ -1,0 +1,1 @@
+export { default as ProfilePage } from "./profile-page";
