--- web/app/src/pages/index.tsx
@@ -1,11 +1,0 @@
-export { default as OpenHouses } from "./admin/open-houses";
-
-export { default as Contacts } from "./contacts";
-export { default as Dashboard } from "./dashboard";
-export { default as OpenHouse } from "./open-house";
-export { default as Projects } from "./projects";
-export { default as Tasks } from "./tasks";
-
-export { default as Login } from "./login";
-export { default as Profile } from "./profile";
-export { default as Register } from "./register";
